% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkProj.R
\name{checkProj}
\alias{checkProj}
\title{checkProj}
\usage{
checkProj(x)
}
\arguments{
\item{x}{\emph{character} vector with a list of paths for the target images or a \emph{RasterLayer} objects.}
}
\value{
A \emph{list} object.
}
\description{
Plots the distribution of images and area per spatial projection.
}
\details{
{For each element in \emph{x}, the function extracts the projection string and estimates the area 
based on the pixel resolution and the number of cells. returns its area. Then, The function identifies the 
unique projections and reports on the number of occurrences and the total area for each of them. The area 
is defined by the maximum extent off all images of a given projection avoiding the over-estimation of the 
area covered by e.g. tiles with a higher temporal coverage.. The output consists of:
\itemize{
 \item{\emph{image.df} - \emph{data.frame} with image-wise reports on projection and area.}
 \item{\emph{proj.df} - \emph{data.frame} with the overal image count and area per unique projection.}
 \item{\emph{count.plot} - ggplot object with the distribution of images per unique projection.}
 \item{\emph{area.plot} - ggplot object with the distribution of area per unique projection.}
}}
}
\keyword{Landsat,}
\keyword{Storage}
