% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateUTM.R
\name{estimateUTM}
\alias{estimateUTM}
\title{estimateUTM}
\usage{
estimateUTM(x = NULL, y = NULL, ext = NULL)
}
\arguments{
\item{x}{Vector of x coordinates.}

\item{y}{Vector of y coordinates.}

\item{ext}{Object of class \emph{Extent}.}
}
\description{
Determines adequate UTM zone for geographical coordinates.
}
\seealso{
\code{\link{satQuery}}
}
\keyword{Landsat,}
\keyword{Storage}
