% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsfeData.R
\name{lsfeData}
\alias{lsfeData}
\title{lsfeData}
\usage{
lsfeData(path)
}
\arguments{
\item{path}{output data path.}
}
\description{
Creates a standardized folder structure.
}
\details{
{The function creates a standardized data structure to ease the transferability of
tools developed within the institute and thus facilitate internal cooperations. The proposed
structure consists of:
\itemize{
\item{\emph{"01_DATA"} - Storage of satellite data and metadata.}
\item{\emph{"02_ANALYSIS"} - Storage of output analysis based on data in "01_DATA".}
\item{\emph{"03_DOCUMENTS"} - General document folder.}
\item{\emph{"04_CODE"} - Folder to store developed code and/or required software.}
\item{\emph{"05_TMP"} - Temporary folder used for data "play".}}}
}
\examples{
\dontrun{

}
}
\seealso{
\code{\link{storeLandsat}}
}
