% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satQuery.R
\name{tileQuery}
\alias{tileQuery}
\title{tileQuery}
\usage{
tileQuery(sensor = NULL, ext = NULL, shp = NULL, tile = NULL,
  dpath = NULL)
}
\arguments{
\item{sensor}{Target sensor. One of \emph{modis}, \emph{landsat} and \emph{sentinel2}.}

\item{ext}{A spatially projected object (e.g. \emph{raster}, \emph{SpatialPolygons}, \emph{SpatialPoints}).}

\item{shp}{Object of class \emph{SpatialPolygonsDataFrame}. Alternative to \☺emph{ext}.}

\item{tile}{Character with name for field in \emph{shp} that contains the tile ID's.}

\item{dpath}{Parent working directory (as provided with \code{\link{lsfeData}}.)}
}
\value{
A \emph{raster} file with the native extention of the input.
}
\description{
Returns Satellite tile for a given location and a list available images on storage.
}
\examples{
\dontrun{

}
}
