% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateUTM.R
\name{estimateUTM}
\alias{estimateUTM}
\title{estimateUTM}
\usage{
estimateUTM(ext)
}
\arguments{
\item{ext}{Object of class \emph{extent} or one from which an extent can be derived. Alternatively, a vector or coordinates can be given.}
}
\description{
Determines adequate UTM zone for geographical coordinates.
}
\seealso{
\code{\link{satQuery}}
}
\keyword{Landsat,}
\keyword{Storage}
